requirements-dev.txt
# Just to emulate source and send data into Event Hub/ Kafka store

kafka-python==2.0.2
requests==2.32.3
azure-eventhub==5.12.1
azure-identity==1.18.0
aiohttp==3.10.5